% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aspect}
\alias{aspect}
\title{Raster of aspect}
\format{
\subsection{\code{aspect}}{

A tif file loaded as a terra spatRaster
}
}
\source{
Derived from the \href{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}{Canadian Digital Elevation Model DEM}
}
\usage{
aspect
}
\description{
Raster of aspect
}
\keyword{datasets}
