% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\docType{data}
\name{basin_dem}
\alias{basin_dem}
\title{Raster of elevation for basin delineation testing}
\format{
A tif file loaded as a terra spatRaster
}
\usage{
basin_dem
}
\description{
Raster of elevation for basin delineation testing
}
\keyword{datasets}
