% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\docType{data}
\name{basin_pts}
\alias{basin_pts}
\title{Points for basin delineation}
\format{
A geopackage file loaded as a terra spatVector
}
\usage{
basin_pts
}
\description{
Points for basin delineation
}
\keyword{datasets}
