% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStreams.R
\name{createStreams}
\alias{createStreams}
\title{Create stream network from DEM}
\usage{
createStreams(DEM, threshold, vector = NULL, save_path = NULL)
}
\arguments{
\item{DEM}{The path to a digital elevation model file, with .tif extension.}

\item{threshold}{The accumulation threshold in DEM cells necessary to start defining a stream.}

\item{vector}{Output file specifications. NULL for no vector file, "gpkg" for a geopackage file, "shp" for a shapefile.}

\item{save_path}{An optional path in which to save the newly created stream network. If left NULL will save it in the same directory as the provided DEM.}
}
\value{
A raster representation of streams and, if requested, a vector representation of streams. Returned as {terra} objects and saved to disk.
}
\description{
Create stream network from DEM
}
