% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permafrost}
\alias{permafrost}
\title{Permafrost data}
\format{
\subsection{\code{permafrost}}{

A data frame with 400 rows and 8 columns
}
}
\source{
Permafrost type classification from central Yukon, with corresponding values derived from the Canadian Digital Elevation Model.
}
\usage{
permafrost
}
\description{
A small data set of permafrost type with correlated terrain attributes
}
\keyword{datasets}
