% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\docType{data}
\name{permafrost_polygons}
\alias{permafrost_polygons}
\title{Polygons of permafrost occurence and type}
\format{
A geopackage file loaded as a terra spatVector.
}
\usage{
permafrost_polygons
}
\description{
Polygons of permafrost occurence and type
}
\keyword{datasets}
