% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solrad}
\alias{solrad}
\title{Raster of solar radiation}
\format{
\subsection{\code{solrad}}{

A tif file loaded as a terra spatRaster
}
}
\source{
Derived from the \href{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}{Canadian Digital Elevation Model DEM} using the \href{https://pro.arcgis.com/en/pro-app/latest/tool-reference/spatial-analyst/area-solar-radiation.htm}{ArcGIS Area Solar Radiation tool}
}
\usage{
solrad
}
\description{
Raster of solar radiation
}
\keyword{datasets}
