% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_load.R
\docType{data}
\name{streams}
\alias{streams}
\title{Lines representing streams}
\format{
A geopackage file loaded as a terra spatVector
}
\usage{
streams
}
\description{
Lines representing streams
}
\keyword{datasets}
