% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{veg}
\alias{veg}
\title{Raster of vegetation types}
\format{
\subsection{\code{veg}}{

A tif file loaded as a terra spatRaster
}
}
\source{
A small subset of the North American Land Cover dataset produced by the \href{http://www.cec.org/north-american-environmental-atlas/land-cover-30m-2020/}{Commission for Environmental Cooperation}, resampled to match cell size of other rasters in this package.
}
\usage{
veg
}
\description{
Raster of vegetation types
}
\keyword{datasets}
