% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basin_dem}
\alias{basin_dem}
\title{Raster of elevation for basin delineation testing}
\format{
\subsection{\code{basin_dem}}{

A tif file loaded as a terra spatRaster
}
}
\source{
Small subset of the \href{https://open.canada.ca/data/en/dataset/7f245e4d-76c2-4caa-951a-45d1d2051333}{Canadian Digital Elevation Model DEM}
}
\usage{
basin_dem
}
\description{
Raster of elevation for basin delineation testing
}
\keyword{datasets}
