% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basin_pts}
\alias{basin_pts}
\title{Points for basin delineation}
\format{
\subsection{\code{basin_pts}}{

A geopackage file loaded as a terra spatVector
}
}
\source{
Created by package developer in ArcGIS Pro
}
\usage{
basin_pts
}
\description{
Points for basin delineation
}
\keyword{datasets}
