% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{permafrost_polygons}
\alias{permafrost_polygons}
\title{Polygons of permafrost occurrence and type}
\format{
\subsection{\code{permafrost_polygons}}{

A geopackage file loaded as a terra spatVector.
}
}
\source{
Permafrost classification polygons created from ground observations and geophysics in central Yukon.
}
\usage{
permafrost_polygons
}
\description{
Polygons of permafrost occurrence and type
}
\keyword{datasets}
