% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{streams}
\alias{streams}
\title{Lines representing streams}
\format{
\subsection{\code{streams}}{

A geopackage file loaded as a terra spatVector
}
}
\source{
A subset of the Yukon CANVEC water flow lines found \href{https://open.yukon.ca/data/datasets/water-flow-50k-canvec}{here}
}
\usage{
streams
}
\description{
Lines representing streams
}
\keyword{datasets}
